Split extracting annotations into a separate file
